updated what i'm learning again
now on to the git and WordPress phase, but I don't like considering it a phase because it makes it sound like it's something that i'll grow out of, but Iidon't want to. it just means that I'll move on to learning something else sometime, which will prob be in my next commit.